Rename tests mods
