+++ node/actors/bft/src/tests/mod.rs
@@ -1,2 +1,2 @@
-mod integration_tests;
-mod unit_tests;
+mod network;
+mod unit;
